+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [Unreleased]
+
+### Fixed
+
+- Fix default value of `userIdAttribute`
+- Fix parameter in `insertLogin()` call
+
 ## [v3.0.7-rciam] - 2021-11-12
 
 ### Added